# app/Http/Controllers/User2Controller.php
<?php

namespace App\Http\Controllers;

use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Hash;

class User2Controller extends Controller
{
    /**
* @OA\Post(
 * path="/api/register",
 * summary="Register a new user",
 * @OA\Parameter(
 * name="name",
 * in="query",
 * description="User's name",
 * required=true,
 * @OA\Schema(type="string")
 * ),
 * @OA\Parameter(
 * name="email",
 * in="query",
 * description="User's email",
 * required=true,
 * @OA\Schema(type="string")
 * ),
 * @OA\Parameter(
 * name="password",
 * in="query",
 * description="User's password",
 * required=true,
 * @OA\Schema(type="string")
 * ),
 * @OA\Response(response="201", description="User registered successfully"),
 * @OA\Response(response="422", description="Validation errors")
 * )
 */
 public function register(Request $request)
 {
 $validatedData = $request->validate([
 'name' => 'required|string|max:255',
 'email' => 'required|string|email|unique:users|max:255',
 'password' => 'required|string|min:8',
 ]);
 $user = User::create([
 'name' => $validatedData['name'],
 'email' => $validatedData['email'],
 'password' => Hash::make($validatedData['password']),
 ]);
 return response()->json(['message' => 'User registered successfully'], 201);
 }
 /**
 * @OA\Get(
 * path="/api/user",
 * summary="Get logged-in user details",
 * @OA\Response(response="200", description="Success"),
 * security={{"bearerAuth":{}}}
 * )
 */
public function getUserDetails(Request $request)
{
$user = $request->user();
return response()->json(['user' => $user], 200);
}

}